GetSubsHist修正輸出型別

+++ WebAPI/Models/Param/Output/OAPI_GetSubsHist.cs
@@ -19,7 +19,7 @@ namespace WebAPI.Models.Param.Output
         public int PeriodPrice { get; set; }
         public double CarWDHours { get; set; }
         public double CarHDHours { get; set; }
-        public double MotoTotalMins { get; set; }
+        public int MotoTotalMins { get; set; }
         public double WDRateForCar { get; set; }
         public double HDRateForCar { get; set; }
         public double WDRateForMoto { get; set; }
